Rewrite the embedded PNG into SVG, for common flags
The switch to SVG happened in 977786c2574ebf05acebd68d8df4c1bc29730f89
and its goal was to allow for hidpi versions of the flags, because they
currently get upscaled and that doesn’t look great.  The plan was to
then progressively convert each flag to a pure SVG version, but it
happened only for very few flags so far.

This reverts the revert in e7062ffe8ec5bb5d19b826fc2007f3efe14757db, but
that was done without any explanation so I’m not sure why it happened.

Some of the flags have slightly different colours, but I based my work
on Wikimedia Commons’ Flag_of_“country name”.svg, so those should be the
most correct versions.  For flags with a different aspect ratio than
3:2, I tried to adhere to the fixup of the current flag (either scale it
up in one direction, or crop it, or extend it), except for the
Macedonian flag where I felt like cropping was better than scaling as
the circle at the middle then became an ellipse otherwise.

I have tested the rendering on my phone and it indeed looks *much*
better.

I only picked the flags happening most often while reloading the front
page, which were simple enough to be smaller as SVG than as PNG.