Speedup expand/collapse translations
The previous implementation recalculated the whole array of displayed
translations when expanding or collapsing, so AngularJS had to render
everything again. Going from expanded state to collapsed state was
fast, but the opposite could be slow with lots of translations.

This new code makes AngularJS render all the translations just once,
keeps them all rendered all the time, and uses ng-hide="" to
conditionally hide some of them when in collapsed state. Much faster.