Fix enabling/disabling of audio
We are sending POST data encoded as x-www-form-urlencoded by default
(as configured in webroot/js/responsive/app.module.js) but does not
allow proper typing. As a result, booleans are sent as strings "true"
and "false" which is a mess.

As far as I understand x-www-form-urlencoded was introduced for backward
compatibility with jQuery code, so we should get rid of it at some point.
When we do, feel free to remove the specific options to $http.post() in
that commit.