Add command to migrate audio files
I changed the location of audio files
in e6cb0f66966b2ea1b95b1fcb75a5760037852478,
so we’ll need this script to move them all to their new locations.

My plan is to hard link the new files so that old files are still
available at no disk space cost to whoever built some script or tool
with the assumption every audio URL can be guessed using the old
naming pattern. To get audio files added after the migration,
audio consumers will have to update because we can’t maintain legacy
names forever.

I intended to only use file linking, but I had to add the "copy" mode
to that command in order to test out because we can’t link()
on tatovm beacuse of a limitation of the virtualbox filesystem.