Reset UI state when refreshing random sentence
The one thing intentionally **not** being reset is the visibility state
of e.g. the translation form, so that users who want to translate
multiple random sentences can do so without having to reopen the form.
But the translation text does get cleared, to avoid accidentally adding
the translation to the wrong sentence.