Show "None" for sentences not having any audio
On the audio import page. This used to be "No" when the column title was
"Has existing audio?", so it makes sense to use "None" with the new column
title "Existing audio". The "None" translatable string happens to be used to
mean "no language" elsewhere in default.pot. I had to useFallback(false) to
prevent CakePHP from using existing "None" translations from default.pot in
case it is not yet translated in this new admin.pot "None".
Different contexts need different translations.