Replace Audio.enabled column by extra table
Keep all enabled audio in the original `audios` table, and
move records we want to disable into a `disabled_audios`
table having the same schema. This design is useful because
99% of the time we only want enabled audios, so we don’t
need to add extra ->where(['enabled' => true]) all over the
existing code. The only place we want disabled audios is
the audio block in the sentence page, for admins, so we
special-case that part only, and keep all the rest as is.

This solves various bugs in listing audio and updating
audio related attributes in Manticore indexes.

This design also allows stats updates of audio to work
out of the box.