Avoid unnecessary connection to wiki db
getArticleTranslations() is making two requests to the wiki
database, one to get $group_id and one to get the article
names. Only the latter was cached, causing an unnecessary
request on every call. To make matters worse, this request,
since it was the very first made to that database,
also caused some others requests for database schema read.

We are having trouble running tatowiki every now and then
(database dead locks and 100% CPU usage) and this commit
should improve the situation.