Turn audio import into background task
This prevents running several imports simultaneously, which
would result in duplicate audio entries and confusing error
messages. As a bonus, this commit allows to import way more
audio at once, that is to say more than what can be
proceeded within the CGI maximum execution time (which
can be very few if the system load happens to be high).

Solves #2833.