Fix unit test
Regression caused by upgrading to CakePHP 3.9:

Failed asserting that two DateTime objects are equal.
--- Expected
+++ Actual
@@ @@
-2022-05-09T21:38:50.043274+0000
+2022-05-09T21:38:50.043330+0000

Because dates now include microseconds, the TimestampBehavior sets
times that are slightly different. The fields 'date' and 'modified'
are set separately, each getting the time at code execution.

Thinking of it, this could lead to having different seconds for both
fields, in the rare case where we enter into the next second right
between both field updates. Since we rely on these fields to be
non-equal in order to treat the record as edited, this could lead
to a minor bug. But I don’t care for now.