Increase Queue maxworkers to 2
When we start running a worker, it actually takes one or two seconds for the worker to effectively start. Because of this, when we try to run the next worker, the previous worker is still working and we cannot run the new one. To prevent this, we need to allow an additional worker to run, just to transition between the old and new worker.