Reword "use new design" setting
I removed the part that says "Note that some features are not yet implemented in this new design..." because it is no longer true. Since we plan to remove the old design, this option will disappear at some point, so I left it in the "experimental" option, but reworded it so that "using the old design" is now considered as experimental.