Turn word count filter into a single method
Even though the code is a bit more complex, it makes
much more sense in terms of API to have a single method.
This will allow to easily add more than one word filter in
the future, such as "word_count > 6 and word_count < 12".

Also added tests for the SentencesSearchForm class.