Make sentences with audio sort order deterministic
Using Audios.modified as a sort key makes the order undefined for audios
having the same modification date. Given than audios are added in bulk,
and the time precision is 1 second, there are lots of such audios.
Besides, sorting on the primary key should give more opportunities for query
optimizations.

Sorting by ID is like sorting by creation date, so it will change the order,
but only for very few audios. Disabling/enabling an audio is the only action
that affects the modification date.