Add option to output all languages
This option will be used on dev.tatoeba.org. The rationale is that whenever
we add a new language on tatoeba.org, a contributor provides a list for that
language, but we also add that language on dev.tatoeba.org, where there are
no sentences, whereas we do need that language to be enabled in search, just
to make sure search works.

By the way, the reason we only include languages having sentences by default
is to lighten RAM consumption of Manticore on TatoVM. I don’t want to change
this default behavior because it could easily mess up with a lot of existing
instances of TatoVM.