Make API 404 responses cleaner
There are cases when NotFoundException is returned, such as
by /audio/{id}/file, when the file cannot be found or read.
In such cases, we don’t want the API to show the HTML error
produced by ErrorHandlerMiddleware, but we want developpers
to be able to see it when debug is enabled.