Fix potential Notice Error
Notice Error: Trying to get property 'username' of non-object in [/home/vagrant/Tatoeba/src/Template/Audio/import.ctp, line 187]

Notice triggered when browsing /audio/import as admin, having uploaded an audio file
named after a sentence number already having an audio which author is not a Tatoeba
user but a user set in the Audio.external field. The Audio.author field is a virtual
field that either brings the associated Tatoeba username or the non-Tatoeba username
depending on which one is set.