Introduce exceptions for search parameters
for sentence language filter.

The current Search model implementation is very loose regarding invalid parameters,
it just ignores them. The reason is that it was implemented primarily for web sentence
search, which goal is to somehow always show search results no matter how wrong
the parameters are (and to eventually, sometimes display a warning message).

Now, in order to reuse the Search model for a search API, we need proper error reporting,
which will be implemented through exceptions.