Simplify API error handling
Display all API errors in json instead of potientially HTML,
so that API consumers get machine-readable errors.

Use Cake HTTP exceptions instead of return $response->withStatus().
This way, when debug is enabled, it helps by showing the location
the exception was thrown from; when debug is disabled,
it shows a more complete message for the consumer.

What’s more, we can now customize the json format of errors.

Note that there is a limitation to this approach of putting an ErrorController.php
file in the prefix routing path: if an exception occurs before routing too early
(such as before routing is done), the error will be shown in HTML anyway. But hey,
this is a limitation of using CakePHP routing mechanism to split execution paths
between website and API.