Rewrite word count filter into a class-based one
Also trying to make testSearchParams() more useful by testing
the post-execution state rather than what methods are internally
called. This is much needed because soon we’ll have nothing but
calls to Search::setFilter() and it will be a mess to know which
filter it is called with and when.