Get rid of $select parameter
Change how filters modify the SELECT clause.
In the returned filter array of [A, B, C] arrays, if B is a string,
it means that that string should be interpreted as a calculated field
for SELECT, and A is the field alias, and C is a boolean to reverse match.

This allows nesting TranslationFilterGroup in itself
to combine arbitrary groups of filters.