Add script to delete stale sessions
Sessions are going to pile up in the database, so we need to clean them up regularly.
PHP has a built-in probability-based garbage collector triggered randomly,
the probability equals to session.gc_probability / session.gc_divisor. The downside
of that approach is that some unlucky request will have to do all the cleanup work
for everybody every now and then. By the way, Debian purposely sets the probability
to zero and provides its own garbage collection script (in /usr/lib/php/sessionclean),
but that script is for the built-in file-based session handler that uses
/var/lib/php/sessions by default. Since we use a custom session handler,
here is a custom cleanup script, which should work for any CakePHP app in theory.