Switch from PHP memcached client to CakePHP's
This is a first step to switch to storing PHP session
data into the database in a seamless way.

Setting php_admin_value[session.save_handler] in php.ini has the effect
of preventing changing session.save_handler at run time. We need to get
rid of that, so that we can use a custom session handler. Next, we’ll
be able to switch the session handler from a simple code update.

The memc.sess.key. prefix seems to be the default prefix used by PHP,
which I discovered using tcpdump.

Refs #3160.