Make sort order deterministic in search results
Add sentence id as secondary sort key, so that the resulst order is
predictable even when the first key is equal.

For example, if two sentences with the same "last modified" time are
returned, when ordered by "last modified", these sentences should
always show up in the same order, and the same search made in reversed
order should always be the opposite of the normal order.

So far it was only the case for the "random" order (which can be
reproducted by providing the same random seed), but now all sort
orders are deterministic.