Switch to keyset pagination in API
Return cursor value along with search results to allow keyset pagination.

Keyset pagination allows to paginate through an arbirary number of results,
not limited by the Manticore "max_matches" limit of 1000 results.

This commit also makes the number of results available in the API response,
but only for the first page. For other pages, it is hidden because keyset
pagination actually turns that value into the "number of results in that page
and subsequent pages" (the number decreases as we go through the pages),
which is rather confusing for a "total".