Speedup unit test execution
The whole test suite is a bit slow, it takes more than a minute to complete,
and minutes to complete when xdebug is enabled, so I am optimizing these slow
tests.

This code change is not very neat, but I don’t think we are going to use
SentenceDerivationShell again in the future, so it’s not very important.
We could delete it entirely at some point.

With this change, on my machine, execution time of
SentenceDerivationShellTest.php drops from about 5s to about 1.3s.