Fix language code rename indexation bug
Prevents a bug that occured during a similar language code
rename in the past, see #2900.

Sentences that are indexed as "having a translation in Old Tupi"
need to be reindexed too, otherwise the search filter "limit to
sentences having a translation in Old Tupi" won’t work any more.
Affected sentences are all the sentences that are directly or
indirectly linked to an Old Tupi sentence.

Here we use the delta reindexation mechanism to get them updated
because the number of sentences Old Tupi is rather small
(563 sentences as of writing).