Lower export script priority
The export script has grown bigger since we are providing
per-language export files (implemented in #2163 and #2360).
While the priority of mysql queries cannot be changed (because
they are performed by the mysql daemon), awk and compression
tasks do bring some extra load on the server that may affect
web clients, especially if the server load happens to be
higher that usual exactly when an export is triggered.