Properly set announcement cookie expiration
Make the "maintenance announcement seen" cookie to expire
exactly when the maintenance starts.

The expiration datetime calculation is based on the server clock since
the maintenance is scheduled on the server. It will be delayed by the
time it takes for the client to load the page, which should be no more
than a few seconds. The delay does not really matter anyway since
the cookie will expire after the maintenance started, not before.