Fix flags color on Browse by language page
SVG flags that include elements not explicitly colored
with fill="..." or style="fill:..." inherit color
from the HTML they are included, thanks to a rule that says:

svg { fill: currentColor }

While this is useful for icons, we want to avoid it for flags.

We have more and more of these "not explicitly colored elements"
because fill="black" is stripped out by SVG optimizers in
tools/generate_flags.sh.