Simplify language icons generation
This changes the generated SVG markup of language icons
that consist of some other icon + a 3-letter code on the side.

Instead of copying the other icon’s markup, we include it using
<use href="..."/>. It is more concise, and always display
correctly even when the linked icon uses different viewBox values.

This way, each svg flag contains no external reference,
so it can be loaded from an HTML <img> (or an SVG <image>).

In addition, we can remove all these extra <symbol> definitions
in the sprites file allflags.svg, so that it each <symbol> is only
declared once and reused across all flags, which makes it smaller
than the sum of all individual flags.