Vectorize language icons based on Berber flag
Berber flag:
- Downloaded from https://commons.wikimedia.org/wiki/File:Berber_flag.svg
- Optimized using https://optimize.svgomg.net/

I added the "centeredL" variant because the result looks better this way.
In the original Berber flag, the ⵣ letter is not perfectly centered,
purposely I assume. By slightly shifting it to the left,
we correct that shift.

The colors are slightly different than the previous flag,
likely because of past color alteration, including in PR #1998.
In the future, let’s keep original colors in the SVG file
and apply any color change using CSS or SVG filters if needed.