Vectorize language icons based on Malaysia flags
Downloaded flag from

  https://commons.wikimedia.org/wiki/File:Flag_of_Malaysia.svg

For Flag_of_Malaysia.svg:

kept the flag 1/2 ratio and optimized the markup a bit.

For zsm.svg:

adjusted size to 2/3 ratio while keeping the star and crescent intact
(the previous flag turned the crescent into an ellipse, I think that’s
ugly) and optimized the markup a bit.