Vectorize Sylheti language icon
This flag uses a 20×20 surface after cropping ISO letters on the right side,
which is different from other flags using a 19×20 surface,
so I added a different template.

It also feature a circle centered on that squared surface,
which is a different position than in ben.svg,
so I added a different base flag benbase.svg.