Make inkscape dependency optional
Only require inkscape when we need it to generate a letter SVG.
This allows contributors to add new flags without inkscape.

This also lessens the probability of messing up with
different inkscape configurations. For example,
`inkscape --export-text-to-path` will produce different
results depending on the local "numeric precision" setting,
which is only accessible from the GUI and not the CLI.