Vectorize language icons based on Guatemala flag
Downloaded from

  https://commons.wikimedia.org/wiki/File:Flag_of_Guatemala.svg

extended top and bottom to ratio 2/3 and resized to 30×20.

To rasterize the central 8×8 block, I edited the flag <svg> with

  viewBox="352 172 256 256" width="256" height="256"

opened the SVG in Firefox and exported it to PNG using an extension.
(Inkscape had problems rendering this SVG correctly.)

Finally, I converted the PNG to JPG using Magick command

  convert guatemala_emblem.png -resize 96x96 -quality 25 guatemala_emblem.jpg