Limit /audio/index pagination and fix order
This limits /audio/index and /audio/index/{lang} endpoints to only list
sentences having the last 1000 audios. We need this because tatoeba.org
now hosts more than 1.2 millions audio recordings. The /audio/index
listing goes on 12000+ pages and /audio/index/eng 8000+ pages (each
with 100 sentences per page). Moreover, we are seeing increasingly
numerous (and sometimes very aggressive) crawlers going through
every possible internal link. This can easily overload the server,
especially when fetching pages greater than 1000, or even 100.

This commit also fixes the audio listing order for sentences having
multiple audio. When a sentence has multiple audio, we only display the
sentence once for clarity. Now, since the sentence list is ordered by
audio id (which is just an efficient way to order by audio creation date),
we should rank a multi-audio sentence according to its most recent audio,
not its oldest. The previous code used DISTINCT to deduplicate the rows;
while it is efficient, it only keeps the last sentence of a group, which
is the oldest. To keep the newest instead, we have to use GROUP BY MAX(id),
which is what this commit does. This GROUP BY aggregation is apparently
more costly than DISTINCT, but since we are now limiting to the first
1000 audio, the aggregation is only performed on 1000 rows,
which is very fast anyway.