Restrict user-configurable sort orders
It is easy to request arbitrary fields using the sort= GET parameter.
CakePHP allows to order by any field by default.

This commits limits the possible fields on which sorting is allowed,
to avoid sorting on non-optimized fields.

This commit also prevents producing a listing of sentences that we were
not able to limit with LimitResultsBehavior. This prevents the potential
bypass of results limit, e.g. by requesting invalid sorts.