Repair /sentence_comments/index listing
Limiting the listing to the latest 1000 items was disabled by
70be3a57a869dc1af4a7297d7da254a80b3df215. Then,
5acebea5fb0e878db3dd19e4865f71ca6b991a38 made the page error
because of the lack of sort order in the $query passed to
findLatest().

This commit makes pagination sort order option go into
the $query parameter of findLatest() of LimitResultsBehavior.php.

To my knowledge, the only way to get the findLatest finder obtain
the sort order specified in the controller’s $this->paginate[]
array is to call $this->paginate() using table+findername as
parameters instead of passing a query object.