Optimize results limitation
It is more efficient to do OFFSET 999 LIMIT 1 than to do LIMIT 1000,
to retrieve the whole result set just to pick the last row.

The main difference is that when the result set is smaller than 1000 rows,
it won’t return any result. In this case, the finder won’t add any WHERE
condition, which is totally fine since there is no need to limit.