Fix returned sentence transcription type in API
The transcription type won’t be correct unless the "Sentences"
containment includes both "lang" and "script" fields.
The code responsible for fetching these fields is actually in
TranscriptionTable::beforeFind(), but it defaults to *all* fields,
whereas we usually only only need "lang" and "script".

Restricting it to only "lang" and "script" revealed that other
fields are needed by TranscriptionTable::findTranscription(),
but that is a special case.

We can handle this special case by using a closure for the contain,
allowing to add extra fields.