Optimize translations retrieval on /sentences API
This add a new (and rather hacky) association type for the sake of
fast retrieval of both direct and indirect translations over a
single SQL query. We had this trick working in CakePHP 2, and some
of that code still remains in TranslationsTable, however it wasn’t
yet possible to use it from a regular ->contain('Translations').

The results end up in a single array that contains both
direct and indirect translations, and each translation entity has
an additional "is_direct" property that can be used for filtering.

I wish to replace the SentencesTable::findFilteredTranslations()
method with this new approach, but because it is new and may has
bugs, I am only adding it to the API for now.