Change showtrans= param behavior in /sentences API
- Rename showtrans= to showtrans:lang=
- Stop returning translations by default
- Only return translations in languages provided in showtrans:lang

Tatoeba supports more than 400 languages and the list of translations
can easily go in the hundreds for some sentences. It does not really
make sense for API consumers to receive so many translations by default,
presumably in languages they are not really interested in.

The /sentences endpoint already requires to specify a sentence language
using lang=, because it does not really make sense to display sentences
in *all* languages. The same reasoning applies for displaying translations.

Besides, we don’t even know if consumers are looking for translations
of for sentences.

The /sentences/{id} endpoint still returns all translations by default.