Add showtrans:is_direct= param to /sentences API
Allow to include all direct translations or all indirect translations
in the response. This parameter can be used alone or combined with
showtrans:lang= to further limit the translations returned.

Now that direct and indirect translations are returned in the same
array (instead of separated in two sub-arrays), it makes sense
to provide this functionality as an easy way to for API consumers
to fetch translations they need without having to do the filtering
themselves.