Add /unstable/audio API endpoint
Simple implementation that allows to browse all audios
by author and language.

There is room for code factorization.

I added failOnRisky="true" to phpunit.xml because of this test
AudioControllerTest::testSearch_ordersResultsByIdAsc().

It needs at least two results to assert that they are ordered by id.
So this test can potentially not perform any assertion, for example
if only one result is returned. In this case, it should be considered
as a failure.