Add API client examples
Also adjust <title> tag contents to show only the last part of the title.
This is because when you have a lot of tabs opened (like me), if they all
start with the same "Tatoeba API..." prefix, all you can see is the prefix
because tab titles are truncated. I noticed that major websites such as
https://developer.mozilla.org/ do not use title prefixes as well.