Limit OFFSET value in SQL requests
When requesting a high-numbered page, such as

  /zh-cn/sentences_lists/show/907/und/und?page=90287

we redirect to the last available page, that is page=100 currently.
However, an expensive SQL request is still executed based on the
number 90287, using OFFSET 902860, just to confirm that the page is
not available and then redirect.

This commit turns the expensive request into a normal one by limiting
the OFFSET value. The redirection behavior shouldn’t be affected.