Redirect sort=created to sort=id in sentence lists
Recently the default sort for lists was changed from "created" to "id"
while introducing a limit in the total number of results, in
2a55b5e2a6480ca30054030c5658ca3ef87d0f95. Even if not available in the UI,
sorting based on "created" was still allowed to avoid breaking many links.

However, it turns out that Tatoeba has a lot of sentences added to lists
at an unknown date, that is, with created = NULL. When browsing such
sentences, the 'latest' finder (defined in LimitResultsBehavior.php) will
fail to find a usable date value to limit the listing, resulting in a
non-limited listing. To ensure the listing is limited,
this commit enforces sorting by id.