Don’t loose comment contents on validation error
When trying to edit an existing comment, a validation error should
bring the user back to the edit form, with the comment input field
filled with the previously sent comment.

CakePHP is designed to do this already, but since we are enforcing
the field value with ['value' => $this->safeForAngular($text)] for
security reasons, it does not work out of the box.