Use Manticore built-in lemmatizer for German
We need this lemmatizer because it works
with non-casefolded tokens, too.

This will allow to index non-casefolded tokens as exact matches,
letting users perform case-sensitive searches. It makes sense for
German because nouns are capitalized and other POS are not,
so one can search for nouns only using '=' search operator.

See #3206.