Reindex sentences when user language level changes
Reindex sentences for the sake of updating the
"sentence is owned by self-proclaimed native speaker" flag.

We could use UpdateAttributes() API to instantly update the new
owner_is_native Manticore index attribute, however my guess is
that nativeness changes will be infrequent and may suddenly
affect a large number of sentences, so it is safer to use the
reindex_flags mechanism instead.