Index sentence owner self-proclaimed nativeness
This will ease searching for sentences and translations
using "is native" filter on both sentences and translations
(currently we cannot do it on translations).

This will also remove the limitation of having to have a
single source language filter selected in order to use the
"is native" sentence filter.

This will also remove the need to work around the maximum
number of values in filters (see issue #2166, closed by
c26b5dbbd9ca37128c769f222db48038fe92c754).

The only drawback is that we need additional work to keep
the Manticore index updated whenever a user’s nativeness
information changes.