Allow audios.{modified,created} to be null
As a matter of fact, we have a lot of audio recordings having an unknown
creation date, because they were imported manually during #1378, with the
introduction of the audios table. These recordings have an audios.created
and audios.modified set to '0000-00-00 00:00:00' as a result of some MariaDB
type casting. Strangely, the casting is still performed when executing e.g.
UPDATE audios SET created = NULL; which only results in a warning.
However, attempting to execute an INSERT with a NULL value results in an
error, which is the root cause of issue #2947 (cannot disable some audio).