Allow translators to use numbers in month selector
Closes #3225.

To help translators acheive proper display of months, this
commit allows them to force numbers in the month selector.

Languages such as Chinese or Japanese do not have names
like January or February. Instead, they use numbers like
"month one" "month two" (一月、二月). It makes more sense
for them to select a number and to put 月 outside of the
dropdown.

However, we already have months as translatable strings,
like "January" is already "一月", so the 月 part ends up
in the dropdown value.