Automatically ban users posting too many links
Since the introduction of the limitation preventing unverified users from
inserting links in their profiles, we are seeing more spam on the Wall.
Basically, content spammers used to put in their profiles get diverted
to the Wall. A large part of it (79%) consists of many links to similarly
exploited websites containing the same spam.

This commit adds a countermeasure to prevent that kind of spam.

The countermeasure is triggered when all these conditions are fulfilled:
- the user is unverified (is_spamdexing = true)
- the user submits or edits a post on the wall containing outbound links
- the post is not already hidden
- the user confirms the content is not for SEO purposes
- the number of outbound links contained reaches a configured threshold

The effect is:
- the content is posted and automatically marked as hidden
- the user role is changed to 'spammer'
- the user is disconnected